new case fixes

+++ app/Admin/Controllers/NewExhibitsCaseModelController.php
@@ -40,7 +40,7 @@ class NewExhibitsCaseModelController extends AdminController
                     $_sus->delete();
                 }
             }
-        } 
+        }
 
 
         $pendingCase = Utils::hasPendingCase(Auth::user());
@@ -150,15 +150,17 @@ class NewExhibitsCaseModelController extends AdminController
 
 
                 $form->select('specimen', 'Specimen')->options(
-                   Specimen::pluck('name','name')
+                    Specimen::pluck('name', 'name')
                 )->rules('required');
 
                 $form->decimal('wildlife_quantity', __('Quantity (in KGs)'));
                 $form->decimal('wildlife_pieces', __('Number of pieces'));
-                $form->text('wildlife_description', __('Description')); 
+                $form->text('wildlife_description', __('Description'));
                 $form->multipleFile('wildlife_attachments', __('Wildlife exhibit(s) attachments files or photos'));
                 $form->divider();
-            });
+            })
+            ->rules('required')
+            ->default('No');
         $form->radio('type_implement', __('Exibit type Implement?'))
             ->options([
                 'Yes' => 'Yes',
@@ -179,7 +181,9 @@ class NewExhibitsCaseModelController extends AdminController
                 $form->text('implement_description', __('Description'));
                 $form->multipleFile('implement_attachments', __('Implements exhibit(s) attachments files or photos'));
                 $form->divider();
-            });
+            })
+            ->rules('required')
+            ->default('No');
 
         $form->radio('type_other', __('Other exhibit types?'))
             ->options([
@@ -191,7 +195,9 @@ class NewExhibitsCaseModelController extends AdminController
                 $form->text('others_description', __('Description for others'));
                 $form->multipleFile('others_attachments', __('Attachments'));
                 $form->divider();
-            });
+            })
+            ->rules('required')
+            ->default('No');
 
         $form->radio('add_another_exhibit', __('Do you want to add another exhibit to this case?'))
             ->options([
@@ -199,11 +205,25 @@ class NewExhibitsCaseModelController extends AdminController
                 'No' => 'No',
             ])->required();
 
-        
+
         // if saving add nothing is added skip to submit //ie no no no
         $form->saved(function (Form $form) {
-            Exhibit::find($form->model()->id)->delete(); //remove the unnecessary exhibit
-        });  
+            if ($form->type_other == 'No' && $form->type_implement == 'No' && $form->type_wildlife == 'No') {
+                Exhibit::find($form->model()->id)->delete(); //remove the unnecessary exhibit
+            }
+            // if($form->add_another_exhibit == 'No'){
+            //     Admin::script('window.location.replace("' . admin_url("new-confirm-case-models/{$form->case_id}/edit") . '");');
+            //     return 'Loading...';
+            // }else {
+            //     Admin::script('window.location.replace("' . admin_url("new-exhibits-case-models/create") . '");');
+            //     return 'Loading...';
+            // }
+            if ($form->add_another_exhibit == 'No') {
+                return redirect(admin_url("new-confirm-case-models/{$form->case_id}/edit"));
+            } else {
+                return redirect(admin_url("new-exhibits-case-models/create"));
+            }
+        });
 
         return $form;
     }
